requirements.txt
alembic==1.13.1
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.1
boto3==1.34.100
botocore==1.34.100
click==8.1.7
dataclasses==0.6
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-Script==2.0.6
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==24.2.1
gevent-websocket==0.10.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.3.5
MarkupSafe==2.1.5
marshmallow==3.21.2
marshmallow-sqlalchemy==1.0.0
packaging==24.0
pluggy==1.5.0
PyJWT==2.8.0
pymongo==4.7.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
s3transfer==0.10.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
urllib3==2.2.1
Werkzeug==3.0.2
wsproto==1.2.0
zope.event==5.0
zope.interface==6.3
